\name{combine_polygons}
\alias{combine_polygons}
\title{Combine polygons}
\usage{
  combine_polygons(sequence_of_polygons, shift = 1)
}
\arguments{
  \item{sequence_of_polygons}{a list of list of matrices.}

  \item{shift}{Optional argument specifying how much the
  successive elements should be horizontally shifted one
  from the other. Default is 0.}
}
\value{
  A list of matrices taken from sequence_of_polygons.
}
\description{
  Combine polygons
}

