\name{compute_ABC}
\alias{compute_ABC}
\title{Pre compute quantities}
\usage{
  compute_ABC(polygon)
}
\arguments{
  \item{polygon}{nx2 matrix representing a polygon;}
}
\value{
  a list of precomputed quantities
}
\description{
  precomputes some quantities specific to the given
  polygon, to optimize subsequent distance computations.
}

