\name{shift_polygon_x}
\alias{shift_polygon_x}
\title{Shift a polygon horizontally}
\usage{
  shift_polygon_x(shift, polygons)
}
\arguments{
  \item{shift}{value of the shift (positive value = to the
  right).}

  \item{polygons}{a list of matrices representing the
  polygons.}
}
\value{
  A list of matrices representing the shifted polygons
}
\description{
  Shift a polygon horizontally
}

