\name{shift_polygon_y}
\alias{shift_polygon_y}
\title{Shift a polygon vertically}
\usage{
shift_polygon_y(shift, polygons)
}
\arguments{
  \item{shift}{value of the shift (positive value = to the
  top).}

  \item{polygons}{a list of matrices representing the
  polygons.}
}
\value{
A list of matrices representing the shifted polygons
}
\description{
Shift a polygon vertically
}

